requirements.txt
hypothesis==3.83.1
Pillow==4.2.1

